+++ user_docs/en/changes.md
@@ -33,6 +33,9 @@ Please open a GitHub issue if your add-on has an issue with updating to the new 
 
 ## 2025.3
 
+This release introduces a new local model-based image description feature.
+You can now use `NVDA+Windows+,` to describe images you encounter.
+
 This release includes improvements to Remote Access, SAPI5 voices, braille and the Add-on Store.
 
 Add-ons in the Add-on Store can now be sorted by minimum/last tested NVDA version and install date.
@@ -47,7 +50,15 @@ There are several bug fixes for SAPI5 voices, including fixes for excessive lead
 eSpeak NG and Unicode CLDR have been updated.
 Localisation data for emojis has been added for Belarusian and Bosnian.
 
+
 ### New Features
+
+* Automated Image Descriptions:
+  * Press `NVDA+Windows+,` to get an AI generated image description. (#18475)
+  * This is generated locally on the device - no information is sent to the internet.
+  * a new command has been added: get an AI generated Image description, assigned to `NVDA+Windows+,`. (#18475)
+  * a new unassigned command is available for quickly opening the settings dialog for local image description. (#18475)
+  * another new unassigned command is available for toggle image captioning. (#18475)
 
 * Add-on Store:
   * Add-ons can be sorted by minimum and last tested NVDA version as well as by installation date. (#18440, #18560, @nvdaes, @CyrilleB79)
@@ -92,6 +103,10 @@ Localisation data for emojis has been added for Belarusian and Bosnian.
 Please refer to [the developer guide](https://download.nvaccess.org/documentation/developerGuide.html#API) for information on NVDA's API deprecation and removal process.
 
 * Component updates:
+  * Updated sphinx to 8.1.3. (#18475)
+  * Updated licensecheck to 2025.1. (#18475)
+  * Introduced onnxruntime 1.19.2 for model inference. (#18475)
+  * Introduced onnx 1.18.0 for generating mock models. (#18475)
   * Updated Pyright to 1.1.403. (#18424)
   * Updated Ruff to 0.12.7. (#18424, #18609)
   * Updated comtypes to 1.4.11. (#18611)